script: Move xml interfaces into xml/
