+++ components/script/dom/mod.rs
@@ -444,11 +444,7 @@ pub(crate) mod workers;
 pub(crate) use self::workers::*;
 pub(crate) mod worklet;
 pub(crate) mod workletglobalscope;
-pub(crate) mod xmldocument;
-pub(crate) mod xmlhttprequest;
-pub(crate) mod xmlhttprequesteventtarget;
-pub(crate) mod xmlhttprequestupload;
-pub(crate) mod xmlserializer;
+pub(crate) mod xml;
 pub(crate) mod xpathevaluator;
 pub(crate) mod xpathexpression;
 pub(crate) mod xpathresult;
