+++ components/script/dom/xml/mod.rs
@@ -1,0 +1,8 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+pub(crate) mod xmldocument;
+pub(crate) mod xmlhttprequest;
+pub(crate) mod xmlhttprequesteventtarget;
+pub(crate) mod xmlhttprequestupload;
+pub(crate) mod xmlserializer;
