+++ components/script/dom/xml/xmlhttprequest.rs
@@ -71,8 +71,8 @@ use crate::dom::servoparser::html::HtmlSerialize;
 use crate::dom::stream::readablestream::ReadableStream;
 use crate::dom::window::Window;
 use crate::dom::workerglobalscope::WorkerGlobalScope;
-use crate::dom::xmlhttprequesteventtarget::XMLHttpRequestEventTarget;
-use crate::dom::xmlhttprequestupload::XMLHttpRequestUpload;
+use crate::dom::xml::xmlhttprequesteventtarget::XMLHttpRequestEventTarget;
+use crate::dom::xml::xmlhttprequestupload::XMLHttpRequestUpload;
 use crate::fetch::{FetchCanceller, RequestWithGlobalScope};
 use crate::mime::{APPLICATION, CHARSET, HTML, MimeExt, TEXT, XML};
 use crate::network_listener::{self, FetchResponseListener, ResourceTimingListener};
