+++ components/script/dom/xml/xmlhttprequestupload.rs
@@ -7,7 +7,7 @@ use dom_struct::dom_struct;
 use crate::dom::bindings::reflector::reflect_dom_object;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::xmlhttprequesteventtarget::XMLHttpRequestEventTarget;
+use crate::dom::xml::xmlhttprequesteventtarget::XMLHttpRequestEventTarget;
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
